int->long (lint complains if you assign RLIM_INFINITY); add Berkeley copyright notice
SCCS-vsn: sys/sys/resource.h 7.2

+++ usr/src/sys/sys/resource.h
@@ -1,9 +1,20 @@
 /*
- * Copyright (c) 1982, 1986 Regents of the University of California.
- * All rights reserved.  The Berkeley software License Agreement
- * specifies the terms and conditions for redistribution.
+ * Copyright (c) 1982, 1986 The Regents of the University of California.
+ * All rights reserved.
  *
- *	@(#)resource.h	7.1 (Berkeley) %G%
+ * Redistribution and use in source and binary forms are permitted
+ * provided that the above copyright notice and this paragraph are
+ * duplicated in all such forms and that any documentation,
+ * advertising materials, and other materials related to such
+ * distribution and use acknowledge that the software was developed
+ * by the University of California, Berkeley.  The name of the
+ * University may not be used to endorse or promote products derived
+ * from this software without specific prior written permission.
+ * THIS SOFTWARE IS PROVIDED ``AS IS'' AND WITHOUT ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ *	@(#)resource.h	7.2 (Berkeley) %G%
  */
 
 /*
@@ -59,6 +70,6 @@ struct	rusage {
 #define	RLIM_INFINITY	0x7fffffff
 
 struct rlimit {
-	int	rlim_cur;		/* current (soft) limit */
-	int	rlim_max;		/* maximum value for rlim_cur */
+	long	rlim_cur;		/* current (soft) limit */
+	long	rlim_max;		/* maximum value for rlim_cur */
 };
